deprecate asSharedSequence in ObservableConvertibleType
